Rewrite make diagnostics to use mod paths
Replaces the make call for mod sources with a new Launch-Make function
responsible for invoking the make cmdlet, grabbing all output, and
rewriting error/warning lines to replace the path to source files so
that they point to the original mod sources instead of the copies of the
sources the build process copies into SDK Development folder.

With this change and the appropriate hooks to your editor to parse an
error line, you can jump directly to errors from a build from within the
editor and it will open the real mod version of the file and not the SDK
copy.